tg_bot_service/requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
certifi==2025.8.3
colorama==0.4.6
frozenlist==1.7.0
h11==0.16.0
hiredis==3.2.1
httpcore==1.0.9
httpx==0.28.1
idna==3.10
loguru==0.7.3
magic-filter==1.0.12
multidict==6.6.4
propcache==0.3.2
pydantic==2.11.9
pydantic_core==2.33.2
redis==6.4.0
sniffio==1.3.1
typing-inspection==0.4.1
typing_extensions==4.15.0
win32_setctime==1.2.0
yarl==1.20.1
python-dotenv==1.0.1